seperated legacy fuciton of params
